pyproject.toml
[project]
name = "llm-knowledge-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.0.3",
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "openai>=2.0.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "spacy>=3.8.7",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
